+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * ``AABB``, ``Circle`` and ``Line`` classes got a ``draw(canvas, color)`` function.
 * ``draw`` function added to ``Vec2``.
 * ``.circle`` getter added to ``AABB`` and ``.aabb`` getter added to ``Sprite``.
+* ``SpriteBatch`` ``.getAnimation`` and ``.getTexture`` now expects ``AssetName`` implementated object.
+* ``.speed`` field added to ``Animation``.
 
 ## 1.0.24